+++ triple_arena/src/chain_iterators.rs
@@ -1,7 +1,84 @@
 //! Iterators for `ChainArena`
 
+use core::marker::PhantomData;
+
 pub use crate::iterators::{CapacityDrain, Drain, Iter, IterMut, Ptrs, Vals, ValsMut};
-use crate::{ChainArena, Link, Ptr};
+use crate::{iterators, Advancer, ChainArena, Link, Ptr};
+
+/// An advancer over the valid `P`s of a `ChainArena`
+pub struct PtrAdvancer<P: Ptr, T> {
+    adv: iterators::PtrAdvancer<P, Link<P, T>>,
+}
+
+impl<P: Ptr, T> Advancer for PtrAdvancer<P, T> {
+    type Collection = ChainArena<P, T>;
+    type Item = P;
+
+    fn advance(&mut self, collection: &Self::Collection) -> Option<Self::Item> {
+        self.adv.advance(&collection.a)
+    }
+}
+
+/// An advancer over the valid `P`s of one chain in a `ChainArena`
+pub struct ChainPtrAdvancer<P: Ptr, T> {
+    // the initial `Ptr` for checking if we are in a cycle
+    init: P,
+    // we ultimately want this in order to provide the extra guarantee that a removal and insertion
+    // into the same spot can't cause the advancer to jump to an unrelated chain
+    ptr: Option<P>,
+    // switch to going in the previous direction
+    switch: bool,
+    _boo: PhantomData<(P, T)>,
+}
+
+impl<P: Ptr, T> Advancer for ChainPtrAdvancer<P, T> {
+    type Collection = ChainArena<P, T>;
+    type Item = P;
+
+    fn advance(&mut self, collection: &Self::Collection) -> Option<Self::Item> {
+        if let Some(ptr) = self.ptr {
+            if self.switch {
+                if let Some(link) = collection.a.get(ptr) {
+                    if let Some(prev) = link.prev() {
+                        self.ptr = Some(prev);
+                    } else {
+                        self.ptr = None;
+                    }
+                    // note how we also get to implicitly check the validity of the original
+                    // `self.ptr` without incurring extra lookups.
+                    Some(ptr)
+                } else {
+                    self.ptr = None;
+                    None
+                }
+            } else if let Some(link) = collection.a.get(ptr) {
+                if let Some(next) = link.next() {
+                    if next == self.init {
+                        // cyclical
+                        self.ptr = None;
+                    } else {
+                        self.ptr = Some(next);
+                    }
+                } else {
+                    self.switch = true;
+                    // `init` was done on first iteration, we need to immediately use the
+                    // previous node to `init`
+                    if let Some(link) = collection.a.get(self.init) {
+                        self.ptr = link.prev();
+                    } else {
+                        self.ptr = None;
+                    }
+                }
+                Some(ptr)
+            } else {
+                self.ptr = None;
+                None
+            }
+        } else {
+            None
+        }
+    }
+}
 
 /// An iterator over `Link<P, &mut T>` in a `ChainArena`
 pub struct ValsLinkMut<'a, P: Ptr, T> {
@@ -21,23 +98,17 @@ impl<'a, P: Ptr, T> Iterator for ValsLinkMut<'a, P, T> {
 /// An iterator for links in a chain in a `ChainArena`
 pub struct IterChain<'a, P: Ptr, T> {
     arena: &'a ChainArena<P, T>,
-    init: P,
-    p: P,
-    switch: bool,
-    stop: bool,
+    adv: ChainPtrAdvancer<P, T>,
 }
 
 impl<'a, P: Ptr, T> Iterator for IterChain<'a, P, T> {
     type Item = (P, &'a Link<P, T>);
 
     fn next(&mut self) -> Option<Self::Item> {
-        if self.stop {
-            None
+        if let Some(p) = self.adv.advance(self.arena) {
+            Some((p, self.arena.get_link(p).unwrap()))
         } else {
-            let p_res = self.p;
-            self.arena
-                .next_chain_ptr(self.init, &mut self.p, &mut self.switch, &mut self.stop);
-            Some((p_res, self.arena.get_link(p_res).unwrap()))
+            None
         }
     }
 }
@@ -98,8 +169,36 @@ impl<P: Ptr, T> FromIterator<T> for ChainArena<P, T> {
 }
 
 /// All the iterators here can return values in arbitrary order, except for
-/// [ChainArena::next_chain_ptr].
+/// [ChainArena::advancer_chain].
 impl<P: Ptr, T> ChainArena<P, T> {
+    /// Advances over every valid `Ptr` in `self`.
+    ///
+    /// Has the same properties as [crate::Arena::advancer]
+    pub fn advancer(&self) -> PtrAdvancer<P, T> {
+        PtrAdvancer {
+            adv: self.a.advancer(),
+        }
+    }
+
+    /// Advances over every valid `Ptr` in the chain that contains `p_init`.
+    /// This does _not_ support invalidating `Ptr`s or changing the interlinks
+    /// of the chain of `p_init` during the loop.
+    ///
+    /// # Warning
+    ///
+    /// This handles cyclical chains, however if links or interlinks of the
+    /// chain that contains `p_init` are invalidated during the loop, or if the
+    /// chain starts as noncyclical and is reconnected to become cyclical during
+    /// the loop, it can lead to an infinite loop.
+    pub fn advancer_chain(&self, p_init: P) -> ChainPtrAdvancer<P, T> {
+        ChainPtrAdvancer {
+            init: p_init,
+            ptr: Some(p_init),
+            switch: false,
+            _boo: PhantomData,
+        }
+    }
+
     /// Iteration over all valid `P`s in the arena
     pub fn ptrs(&self) -> Ptrs<P, Link<P, T>> {
         self.a.ptrs()
@@ -123,20 +222,11 @@ impl<P: Ptr, T> ChainArena<P, T> {
     }
 
     /// Iteration over `(P, &Link<P, T>)` tuples corresponding to all
-    /// links in the chain that `p` is connected to. This starts at the link
-    /// pointed to by `p` and iterates along the `next` direction until it
-    /// reaches the end of the chain (or reaches `p` again if the chain is
-    /// cyclical), then it iterates in the reverse direction starting from the
-    /// link before `p` if it exists. If `p` was invalid, this
-    /// iterator returns `None`.
-    pub fn iter_chain(&self, p: P) -> IterChain<P, T> {
-        IterChain {
-            init: p,
-            p,
-            switch: false,
-            stop: !self.contains(p),
-            arena: self,
-        }
+    /// links in the chain that `p_init` is connected to, according to the order
+    /// of [ChainArena::advancer_chain]
+    pub fn iter_chain(&self, p_init: P) -> IterChain<P, T> {
+        let adv = self.advancer_chain(p_init);
+        IterChain { arena: self, adv }
     }
 
     /// Mutable iteration over `(P, Link<P, &mut T>)` tuples
@@ -154,98 +244,5 @@ impl<P: Ptr, T> ChainArena<P, T> {
     /// Same as [crate::Arena::capacity_drain]
     pub fn capacity_drain(self) -> CapacityDrain<P, Link<P, T>> {
         self.a.capacity_drain()
-    }
-
-    /// Same as [crate::Arena::first_ptr]
-    #[must_use]
-    pub fn first_ptr(&self) -> (P, bool) {
-        self.a.first_ptr()
-    }
-
-    /// Same as [crate::Arena::next_ptr]
-    pub fn next_ptr(&self, p: &mut P, b: &mut bool) {
-        self.a.next_ptr(p, b);
-    }
-
-    /// Same as [crate::Arena::next_ptr] except that this only iterates over
-    /// elements of the chain that `p` is connected to. This does _not_ have
-    /// a corresponding `first_chain_ptr` function, see the below
-    /// recommended structure for calling this.
-    ///
-    /// ```text
-    /// let init = ...;
-    /// let mut p = init;
-    /// let mut switch = false;
-    /// let mut stop = !arena.contains(init);
-    /// loop {
-    ///     if stop {
-    ///         break
-    ///     }
-    ///
-    ///     // use `p` here, but be aware that the removal or insertion of
-    ///     // elements within the chain of `init` is not supported
-    ///
-    ///     arena.next_chain_ptr(init, &mut p, &mut switch, &mut stop);
-    /// }
-    /// ```
-    ///
-    /// For the details on how this works (as long as you adhere to the format
-    /// above, you do not have to worry about this): On the first iteration of
-    /// "use `p` here", `p == init` is used. `next_chain_ptr` will then check
-    /// for the next link. If the next link is equal to `init`, then we know we
-    /// are in a cyclical chain and `stop` is set. If there is a next link, `p`
-    /// is set to that next link and iteration progresses until the end of the
-    /// chain. When the end of the chain is reached and there is no next link,
-    /// `switch` is set and `p` is set to the link previous to `init` (or sets
-    /// `stop` if there is no previous link). When `switch` is set previous
-    /// links are similarly followed until the end of the chain, when `stop` is
-    /// set.
-    ///
-    /// If `p` is invalid, `stop` is set, but should only ever be the case if
-    /// the initial `init` was invalid.
-    pub fn next_chain_ptr(&self, init: P, p: &mut P, switch: &mut bool, stop: &mut bool) {
-        if *switch {
-            let prev = if let Some(link) = self.a.get(*p) {
-                if let Some(prev) = link.prev() {
-                    prev
-                } else {
-                    *stop = true;
-                    return
-                }
-            } else {
-                *stop = true;
-                return
-            };
-            *p = prev;
-        } else {
-            let next = if let Some(link) = self.a.get(*p) {
-                if let Some(next) = link.next() {
-                    next
-                } else {
-                    *switch = true;
-                    // `init` was done on first iteration, we need to immediately use the previous
-                    // node to `init`
-                    if let Some(link) = self.a.get(init) {
-                        if let Some(prev) = link.prev() {
-                            *p = prev;
-                        } else {
-                            *stop = true;
-                        }
-                    } else {
-                        *stop = true;
-                    }
-                    return
-                }
-            } else {
-                *stop = true;
-                return
-            };
-            if next == init {
-                // cyclical
-                *stop = true;
-                return
-            }
-            *p = next;
-        }
     }
 }